Assingment #2
Create a Java program that successfully displays in the execution command shell, an ASCII character graphic of the room and painting information as it is specified below. All of the numeric values must be calculated in your code and displayed in the output.
Learn to declare and assign primitive type variables, cast primitive types, and use the java.lang.Math library. Continue becoming acquainted with the process of developing and testing executable Java code; using of an editor, compiler, and runtime environment as programming tools; and deploying the developed capability on a remote host.